break out a bunch of stack tasks into files
