+++ lib/stax.rb
@@ -10,7 +10,10 @@ require 'stax/staxfile'
 require 'stax/base'
 require 'stax/cli'
 require 'stax/stack'
-require 'stax/cfn'
+require 'stax/stack/cfn'
+require 'stax/stack/crud'
+require 'stax/stack/parameters'
+require 'stax/stack/outputs'
 require 'stax/subcommand'
 
 require 'stax/git'
